feat(channels): create and integrate Channels component for social links in Contact, SignIn, and SignUp pages
